src/test/java/servlet/SimpleServletTest.java
package servlet;

import static org.junit.Assert.assertTrue;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.PrintWriter;
import java.io.StringWriter;
import org.junit.Test;
import server.controller.SimpleServlet;

public class SimpleServletTest {

  @Test
  public void simpleDOGETtest() throws Exception {
    HttpServletRequest request = mock(HttpServletRequest.class);
    HttpServletResponse response = mock(HttpServletResponse.class);

    StringWriter stringWriter = new StringWriter();
    PrintWriter writer = new PrintWriter(stringWriter);
    when(response.getWriter()).thenReturn(writer);

    new SimpleServlet().service(request, response);

    writer.flush();
    assertTrue(stringWriter.toString().contains("Hello, World!"));
  }

}